feat: create rule from transaction
Add a "Create Rule" shortcut to the transactions UI, accessible from
two places: the bulk action bar (when one transaction is selected) and
the transaction edit dialog footer.

Opening it pre-fills the rule dialog with the transaction's description
as a "contains" condition and its category as the default action, so
the user only needs to name the rule and save.
