+++ frontend/src/locales/en.json
@@ -233,6 +233,7 @@
     "noCategory": "No category",
     "addManual": "Add Transaction",
     "duplicate": "Duplicate",
+    "createRule": "Create Rule",
     "saveAndNew": "Save & New",
     "saveAndDuplicate": "Save & Duplicate",
     "moreSaveOptions": "More save options",
