+++ frontend/src/locales/pt-BR.json
@@ -233,6 +233,7 @@
     "noCategory": "Sem categoria",
     "addManual": "Adicionar Transação",
     "duplicate": "Duplicar",
+    "createRule": "Criar Regra",
     "saveAndNew": "Salvar e novo",
     "saveAndDuplicate": "Salvar e duplicar",
     "moreSaveOptions": "Mais opções de salvar",
